removing old flow

+++ src/supervisor/task.cpp
@@ -4,15 +4,18 @@
 
 
 Task::Task(Callback callback):suspended(false),task_callback(callback),priority(P_LOW){
-	ptr_value=(int)this;
+	ptr_value=(uint16_t)this;
+	//uint16_t ptr = (uint16_t) &value; 
 }
 
 Task::Task(Callback callback,bool suspend):suspended(suspend),task_callback(callback),priority(P_LOW){
-	ptr_value=(int)this;
+	ptr_value=(uint16_t)this;
+	//uint16_t ptr = (uint16_t) &value; 
 }
 
 Task::Task():suspended(false),priority(P_LOW){
-	ptr_value=(int)this;
+	ptr_value=(uint16_t)this;
+	//uint16_t ptr = (uint16_t) &value; 
 }
 
 Task::~Task(){
